Use intptr_t in numpy loop dimensions and steps.

+++ libgumath/apply.c
@@ -67,8 +67,8 @@ gm_apply(const gm_kernel_t *kernel, xnd_t stack[], int outer_dims,
         const int dims_size = outer_dims + sum_inner;
         const int steps_size = nargs * outer_dims + sum_inner;
         ALLOCA(char *, args, nargs);
-        ALLOCA(int64_t, dimensions, dims_size);
-        ALLOCA(int64_t, steps, steps_size);
+        ALLOCA(intptr_t, dimensions, dims_size);
+        ALLOCA(intptr_t, steps, steps_size);
 
         if (gm_np_convert_xnd(args, nargs,
                               dimensions, dims_size,
